pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aica"
version = "0.1.0"
description = "AICA - An MGX-like AI coding assistant using MetaGPT"
authors = [
    { name = "Dan Rapp", email = "rappdw@gmail.com" }
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "boto3>=1.34.0",
    "openai>=1.0.0",
    "pydantic>=2.5.3",
    "typer>=0.9.0",
    "rich>=13.6.0",
    "pyyaml>=6.0.1",
    "aiohttp>=3.8.6",
    "tenacity>=8.2.3",
    "tiktoken>=0.7.0",
    "loguru>=0.6.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "ruff>=0.1.0",
    "flake8>=7.0.0",
    "mypy>=1.0.0"
]

[project.scripts]
aica = "aica.cli:app"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
